requirements.txt
appdirs==1.4.3
attrs==18.2.0
biopython==1.72
certifi==2018.10.15
chardet==3.0.4
ConfigArgParse==0.13.0
cycler==0.10.0
Cython==0.29.1
datrie==0.7.1
Distance==0.1.3
docutils==0.14
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
idna==2.7
interlap==0.2.6
jsonschema==3.0.0a3
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
pyfaidx==0.5.5.2
pyparsing==2.3.0
pyrsistent==0.14.7
pysam==0.13
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
ratelimiter==1.2.0.post0
regex==2018.11.22
requests==2.20.1
scipy==1.1.0
six==1.11.0
smmap2==2.0.5
snakemake==5.3.0
umi-tools==0.5.5
urllib3==1.24.1
wrapt==1.10.11
